Rule became a large class, with lots of materialisation logic. Create a new Materialiser to hold these static methods. Rename the Materialiser used in tests to better indicate it's use as ForwardChainingMaterialiser.
